+++ doc/source/changelog.rst
@@ -4,6 +4,13 @@ Changelog
 All notable changes to this project will be documented in this file.  This
 project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
+Version 1.1.1
+-------------
+
+Internal:
+
+  * An integer type error has been fixed.
+
 Version 1.1.0
 -------------